+++ app/web/__init__.py
@@ -4,6 +4,7 @@ from .leaves.routes import leave_ns
 from .cities.routes import city_ns
 from .objects.routes import object_ns
 from .object_statuses.routes import object_status_ns
+from .project_schedules.routes import project_schedule_ns
 from .projects.routes import project_ns
 from .users.routes import user_ns
 from .shift_reports.routes import shift_report_details_ns, shift_report_ns
@@ -23,6 +24,7 @@ def register_namespaces(api: Api):
     api.add_namespace(city_ns)
     api.add_namespace(object_ns)
     api.add_namespace(object_status_ns)
+    api.add_namespace(project_schedule_ns)
     api.add_namespace(project_ns)
     api.add_namespace(user_ns)
     api.add_namespace(shift_report_ns)